Remove OWASP scan on every PR
We may want to bring this back someday. This scanner seems to operate by
making a Github issue with any scan discrepancies. That seems fine to
run daily, but on every PR it may be noisy if it is creating separate
issues for those.

I read through the warnings, and none seemed to be really that
significant. I also downloaded the local ZAP program in hopes of
replicating the warnings, but got an (almost entirely) different set of
warnings back.

For now, until we understand this tool better, I'm inclined to disable
it.