Bring in UswdsFormBuilder
This class was written for another Nava project, and can be useful for
us to build off in order to produce USWDS form styling. In order to get
the tests running, I had to bring in Capybara, which seems okay since we
will probably want to use it soon for integration testing anyway.