Mount /rails/tmp as temp directory
Puma writes its pid to a temporary pidfile in ./tmp/pids by default,
however, this was a problem because we were mounting the whole
filesystem as readonly.

To allow for creation of this pid file and whatever else Rails needs
that tmp directory for, let's mount it in as an ephemeral volume from
Fargate.

This commit also includes some README changes to reflect my experience
getting terraform and docker running on my laptop.