Remove credentials.yml.enc in favor of env vars (#38)
* Fix rubocop style issues

This commit runs `rubocop --autocorrect` to fix the style issues.

* Remove encrypted credentials file

As this is a public repo, we won't be storing credentials encrypted in
this file. Instead, we will use environment variables.

* Remove cloud.gov Redis configuration

This appeared to be attempting to configure Redis based off an
environment variable called VCAP_SERVICES with some kind of JSON blob of
configuration. This appears to be a Cloud Foundry concept that we aren't
using.