Set default host for emails based on a DOMAIN_NAME envvar
When generating links in email, we need to specify a
`default_url_options[:host]` configuration so Rails knows what the
hostname will be. We have been doing this in development (with
"localhost"); let's do the same thing in production (with whatever the
value of the service's `domain_name` variable is).