Add PDF test rendering endpoint
I wanted to test PDF rendering without having to go through the whole
flow. So this adds a development-only endpoint to test PDF rendering, at
/health/test_rendering.pdf.

This may be useful in production environments, too, but runs a DDOS
risk, so that's why I've restricted its access for now. We could figure
out a more secure approach to open it up later.