Enable "ECS Exec" for debugging in AWS ECS
I want to look at our database in RDS to debug some strange behavior.
Thankfully, we already have [the configuration in place][1] to be able
to use ECS Exec to get me a shell in there. This commit enables the
configuration, and adds a helper script, so we can get a Rails console
in the demo environment running via:

```bash
bin/ecs-console
```

Also fixed/changed:
* Install `ngrok` correctly via Brew
* Install `awscli` and the necessary `session-manager-plugin`
* Revert `readonlyRootFilesystem` configuration to be dependent on
  whether command execution is enabled

[1]: https://github.com/DSACMS/iv-cbv-payroll/blob/main/docs/infra/service-command-execution.md