Use CBV invite as higher priority than whatever's in session
This is probably only an issue for development, while we have a CBV flow
in our session for development and then want to later test following a
CBV invite via tokenized link. The current behavior is that if there is
a CBV flow in the session, it will be prioritized over the tokenized
link.

But this seems a bit mistaken. If two people are applying for SNAP
one-after-the-other on the same device, we will want the second link to
work despite the presence of the first flow_id in the session.