Update PinwheelWebhookManager to not recreate the same webhook
When the Rails server is reloading, the Pinwheel webhook will be the
same URL since Ngrok hasn't restarted. It's inefficient to recreate the
webhook every time.

This commit updates the PinwheelWebhookManager to look for an existing
webhook with the same URL and events and don't do anything if that
already exists.