Update task_executor IAM policy before running migrations
When adding an environment variable to AWS SSM, the task executor will
need access to read the environment variable in order to start up a
container. This is normally handled later in the deploy, when the
`bin/deploy-release.sh` script runs `make infra-update-app-service`. But
by then, the migrations container will already have failed.

This commit updates the targets for the pre-migration `terraform` run to
also update the `aws_iam_role_policy.task_executor` resource so that it
will have the permissions it needs to start up the migrations.