Add terraform config for CBV production environment
This commit is the result of:
1. Adding AWS access keys to my local ~/.aws/credentials file (under a
   new section for "prod")
2. Running:
   ```
   AWS_PROFILE=prod make infra-set-up-account ACCOUNT_NAME=nava-ffs-prod
   AWS_PROFILE=prod make infra-configure-app-service APP_NAME=app ENVIRONMENT=prod
   AWS_PROFILE=prod make infra-configure-app-build-repository APP_NAME=app
   AWS_PROFILE=prod make infra-configure-app-database APP_NAME=app ENVIRONMENT=prod
   ```
   and their corresponding `update` make targets that actually create the AWS
   resources.
3. Reviewing and making minor updates to the deploy documentation

One note is that this changes the shared build repository to be hosted
in production. It appears the infra template is architected around a
single shared container image registry, so, keeping that in production
probably makes more sense than having it in dev(/demo) and having
production pull from there.