Fix production deploy by pushing built images to prod registry
Production deploys are currently failing due to the task not being able
to [pull the built image from the registry][1]. Upon further inspection,
this is probably because there is a mismatch between the configuration
for the "shared" AWS environment (which points to the "nava-ffs" dev
account and will be used by GH Actions when [building and pushing the
image][2]) and the actual [configuration of the "shared" Terraform
backend][3] (which points to the production account).

This commit harmonizes these environments in a hope that it just works.
I can't test this, since it's Github Actions, so I'm going to rely on
merging and seeing how it goes after that.

I also removed "staging" config for now, since we don't intend to have a
staging environment right now. I've left it commented-out so we can
quickly retrieve it if we see the need for it later.

[1]: https://github.com/DSACMS/iv-cbv-payroll/actions/runs/10171383905/job/28134142130
[2]: https://github.com/DSACMS/iv-cbv-payroll/blob/82a4f2b2/.github/workflows/build-and-publish.yml#L45-L52
[3]: https://github.com/DSACMS/iv-cbv-payroll/blob/82a4f2b2/infra/app/build-repository/shared.s3.tfbackend