Fix demo deploys by pulling image from correct account_id
Deploys to our demo (dev) environment were broken because they were
trying to pull the built image from the wrong AWS account (dev instead
of prod). This is because the service module was using a
"data.aws_ecr_repository" module to load the repository by name, despite
the fact that the ECR repository was in another account.

The fix for this is to pass in the "image_repository_account_id" to the
service module. We determine the value when processing the app-config by
calling the bin/account-ids-by-name.sh script, and looking for the
account ID for the account that contains the "shared" environment.

This commit also updates the publish-release.sh script. Although this
script worked (as the Github Action properly ran it in the shared
environment), I thought it would be good to standardize how the image
repository values were accessed.