Move Pinwheel API token into site-config.yml
This will allow us to have sites using different Pinwheel
configurations. The intent here is that we will be able to have a
"sandbox" site that is always using the Pinwheel sandbox while we can
have NYC and MA in the Pinwheel "development" environment so they can
access real data.

This necessitates always having a "site_id" present when instantiating a
PinwheelService object, so that it can pull the correct API token from
the site configuration. Every place we do it except one, we have access
to a `cbv_flow` object, so it works. (The exception is the dev webhooks
manager, which I updated to only create webhooks for the "sandbox"
site.)

There is one gotcha in the Pinwheel webhook events controller: we
previously validated the webhook signature *before* loading the CbvFlow
model (good!) However, now we need to load the CbvFlow object first in
order to have the API token to use to validate the webhook signature. To
prevent timing attacks that would allow enumeration of the
`pinwheel_end_user_id` field, if the CbvFlow object is not found, we
still pretend to validate the webhook signature using a dummy API token
so it takes approximately the same duration.