Update production domain to snap-income-pilot.com
It's really this easy! I just had to run the `make` commands to update
the network and the service, after deleting the duplicate Route53 zone
that was automatically created after registering the domain.

(Also fixed a bug in accessing the `account-ids-by-name` script that was
preventing it from being run in the `make infra-update-network` target.)