Fix employer search inadvertently using CSP
The Pinwheel employer search will not work when a CSP is in effect,
since Pinwheel uses inline <style> tags. The CSP is disabled in the
controller, however, when the employer search page is requested via
Hotwire/Turbo, the headers that disable CSP don't apply to the root html
document.

So, disabling turbo on the previous page form will do the trick -
forcing the browser to do a full load of the next page with the proper
CSP.