Redirect to /cbv/expired_invitations after CBV flow is complete
After the CBV flow is complete, we want to prevent the user from going
back and changing anything. The easiest way to do this is to clear out
the `cbv_flow_id` from the session on the confirmation page, and then
strengthening our enforcement that every page in the CBV flow follows
from an invite.

A CBV flow is complete if it has a confirmation_code assigned.