Add initial implementation of Data Redaction
This is based off a technique I used in a previous project: rather than
deleting data, we will replace the values for all sensitive fields with
placeholder values (e.g. "REDACTED") so that foreign key associations
are maintained and we can still have residual analytics on non-sensitive
fields.

This commit implements a Redactable module that can be included into any
model to configure its redaction settings. A `DataRetentionService` then
calls the redaction methods.

Future commits will keep track of this redaction as well as implement
more fine-grained rules for CbvFlowInvitation so that they are not
immediately deleted after 7 days.