Add support to infra template for cron jobs
We want to run a job every hour to redact data older than a threshold (7
days). This is pretty similar to the support the infra template already
has for file-upload-triggered jobs, so I've put the configuration right
alongside those.

The task is scheduled with EventBridge Scheduler, which runs a
serverless cron. The Scheduler itself has a different service identifier
than the event manager used for the file-upload jobs, so I created a
separate IAM user and policy to allow it to AssumeRole.