Add `site_translation` helper for site-specific copy
We've needed this ever since the initial concept of site config: a way
to differentiate copy based on which site is being viewed. Rather than
adding all of these strings to the site-config.yml, let's create a
helper method that looks for the key dynamically based on the current
site.

This commit removes the site config for `agency_help_link`,
`learn_more_link_text` and `learn_more_link_url` by replacing them with
some site-specific translations.