Fix mailer previews so they work with FactoryBot
They were returning an error because when FactoryBot is used outside of
rspec, you have to manually call `find_definitions`. This may be
overkill, but I created a BaseMailerPreview class so we can get that
behavior automatically in other preview clasess.

This also adds a new mailer preview for DTA vs NYC so we can try viewing
them both.