Redirect to homepage after timed out employer search
If the session times out right before the user searches for an employer,
the homepage redirect will be handled by turbo, which is expecting to
find a <turbo-frame id="employers"> in the response. Since the homepage
doesn't have such a frame, it displays a "Content missing" error message
and the user is stuck.

We need to catch the failure event for this page and force a top-level
page redirect.