Revert "Use "Strict" Set-Cookie SameSite value"
This reverts commit 5b788c91a1a1fb8192d805d017cf117284f8dd64.

Reading the definition of "SameSite=Strict" a bit closer, it says

> If a request originates from a different domain or scheme (even with
> the same domain), no cookies with the SameSite=Strict attribute are
> sent.

This was breaking SSO because the session cookie wasn't being sent to
the OmniAuth callbacks controller, since the navigation request
originates from a different domain (Microsoft's SSO page).

We don't actually need Set-Cookie=Strict for security reasons, it's only
a nice-to-have.