Refactor `site_translation` helper again to use `current_site` method
Unfortunately my refactor in e485b4d13, where I replaced the
`current_site` method with a `@current_site` instance variable, broke
the caseworker invite page because the `current_site` method which sets
the `@current_site` variable was never set.

Since this is very weird for a function to rely on the definition of an
instance variable, let's go back to it just calling the `current_site`
method and make it explicit in the documentation string that each
controller/mailer has to define a `current_site` method.

This fixes the missing DTA agency name in the caseworker dashboard.