Add more metadata when tracking PDF share event
We need to track a few fields in order to give us metrics on how mayn
accounts are linked, as well as the number of paystubs and account
descriptions.

This logic gets me worried for bugs, since it's in a critical place, so
I'm putting it in a rescue block so it doesn't break the sending of the
actual documents.