Surround most text in a "usa-prose" component
The typography on our pages doesn't match the designs in general:
everything fees cramped with a short line-height and different
font-size. Upon further inspection of the USWDS page templates, most
body content is in a "Prose" component, which makes the text a bit more
readable.

This commit surrounds all the text in a "usa-prose" component.

It also updates the styling of the <label> with the consent, since there
was an extra large bottom margin due to the <p>.

I didn't touch the payment details page, since Matt is concurrently
refactoring the whole thing.