Track Pinwheel error events in NewRelic
We can get a callback from the Pinwheel modal whenever there is an error
that blocks a user from signing in. This commit adds a rudimentary error
handler to forward that event as a custom "PageAction" event to
NewRelic.

Ideally, this would allow us to correlate to a CBV flow instance, and
also be resilient to any ad blockers (uBlock blocks NewRelic metrics),
but this is a good first step.