Fix `employer_name` display on summary page/PDF
On the payment details page, we were rendering the `employer_name` based
off the `employments` API endpoint, which seems to contain data more
often than the individual paystub does.

This commit updates the summary page to also pull the employer name from
the employments hash.

Along the way, it fixes a bug that prevented rendering the PDF for my
Nava data due to a missing `employer_phone_number` hash.