Add terraform to subscribe to email delivery/failure events
This adds terraform code to get us most of the way towards processing
the delivery status of every email: it sets up delivery event
notifications for any successfully delivered email, or any email that is
not deliverable due to bounce.

The remaining work is to actually subscribe the application to receive
these events as HTTP webhooks, either directly via SNS or by creating a
Lambda to send the events. Until that work is done, we can get
visibility into email deliverability (for debugging purposes) by
subscribing our personal email addresses to it. (Again, only for
debugging purposes until we have actual real data.)