Fix /:site/invitations to be a 404 instead of 500
This error has happened a couple times. I couldn't reproduce why, so I'm
assuming that NYC staff just forgot the "/new" at the end of the link
and typed it in manually. To handle this, let's make it be a 404 rather
than 500 error.

While testing, I also cleaned up the parameters used on these redirects
- we don't have an invite secret anymore.