1577: Fix error that happens during PDF rendering
The issue was that, when precompiling assets during the Docker build,
the DOMAIN_NAME environment variable is not set, leading to incomplete
URLs being referenced by the stylesheets built by the asset pipeline,
for example:

  background-image: url(https:/assets/@uswds/uswds/dist/img/file[...].svg)

Since the same asset build is used for both demo and production, we
can't have a domain name in there. So, let's unset the
`config.asset_host`, so that no domain name is included during
compilation, and instead set the fine-grained
`config.action_mailer.action_host`, which is only used in the mailers.