Add script to ease upgrading template-infra
The template-infra repo contains a script to upgrade itself, but this
script didn't work for me because it applies patches as an
all-or-nothing operation, and of course we have made modifications to
our terraform that will cause the patch application to fail.

To make it easier on myself, I wrote a script that automatically looks
for new released versions of `template-infra` and steps through
upgrading them one version at a time. Each version is applied by
applying the individual commits, and the pool soul doing the upgrade
will have to resolve any conflicts commit-by-commit.

Using this script, I was able to upgrade 3 minor versions and 4 patch
versions in about 3 hours of work. I still did have to clean up the git
history after this, and make sure everything still works, so the script
could certainly be improved, but alas, I need to timebox this due to
launch concerns.