1630: Fix MA MOVEit transmission by installing gpgme
We need to install `libgpgme11` in order for the GPGME rubygem to be
able to encrypt the files.

This commit also upgrades git (as it seems to be necessary for the build
container) and alphabetizes the list of dependencies so it's easier to
determine where to add a new one in the future.