1646: Invalidate sessions on logout (addresses MA security finding)
This may be too clever to last long-term, but, to prevent against reuse
of session after logout, this commit adds logic to maintain a JSON
object of invalidated session IDs in the `users` table.

When logging out, the session ID is stored in that JSON object. If the
user attempts to reuse the old session, we will not accept the login.

This also enforces server-side session timeouts of 30 minutes, to
prevent re-use of old sessions.