Fix `invalidated_session_ids` system error when logging out twice
The "before_logout" hook can be called with a `nil` user if the user has
already logged out (or their session timed out, presumably). In this
case, we can't do anything to invalidate the session because Devise
won't tell us which user it was related to. That's fine - we don't need
to invalidate the session, as Devise will already consider the session
timed out.

This commit fixes the error that comes up with `user` is nil in this
case.