Implement E2E tests with Capybara
This is a spike of a couple hours' work to show what E2E tests with
Capybara could look like. This implements "Task 1" of the tech spec
discussed on 9/27.

These tests won't run during CI, but maybe they will be useful in our
manual testing. Future plans are to stub out Pinwheel and then enable
these in CI.