Fix RDS parameter group upgrading to create_before_destroy
When upgrading to v0.10.0, we hit an undocumented breaking change - the
change of the RDS parameter group name necessitates Terraform recreate
the resource. But, the parameter group cannot be destroyed while the
running database is currently using it.

By creating a new parameter group first (before destroying the old one),
we can handle upgrades to the parameter group without causing breaking
changes.