FFS-1661: prevent using the back button after flow is complete (#308)
Before these changes, if you hit back on the success page, you could load a cached page and successfully go back (which is bad). After these changes, when you try this, you will not load a cached page and you will not be allowed to go back (which is good).