Add email open/deliver/bounce events to NewRelic
This commit updates our email terraform to forward email delivery events
to NewRelic via EventBridge. (We don't need an SNS topic, as a result.)

The Email events are in NewRelic as custom events named "AWSSESEvent",
broken down with attributes for the SES event type and the message ID,
so we can make a funnel.

Unfortunately, we won't be able to tie these to the CBV Flow or
invitation just yet, but perhaps we can add a custom email header.

If we want to tie these in with CBV Flow or invitations, we could
instead send these to our application as HTTP webhooks using
EventBridge (by duplicating the code here).