Fix data redaction for Cbv Flows without invitations (#334)
We have a number of these in the database from early in the application:
Cbv Flows that didn't originate from an invitation. Currently, the data
redaction is failing for these records because we use the invitation's
expiration date as the threshold for when to redact the data.

We'll need to handle this situation properly given we're planning to
allow invitation-less CBV for real in the near-future, so let's update
the logic here so we redact Cbv Flows based on either:
* The invitation's expriation + 7 days (if invited)
* The Cbv Flow's most recent update + 7 days (if not invited)

By using the update timestamp, we can try our best to not redact an
active Cbv Flow that someone is actively completing (i.e. if they
started the flow and then kept their browser active for 7 days).