Rewrite PR template to emphasize Acceptance Testing plan
This change is meant as a proposal to encourage a bit more team emphasis
on Acceptance Testing plan when creating new PRs. By selecting one of
the three main acceptance testing methods, the code reviewers will share
an understanding with the PR's author about future steps necessary
before the code is deployed to production.