FFS-2340: Create `WebhookEvents` to track sync success/failure
We currently use Pinwheel-specific columns on `PayrollAccount` to track
the status of the sync: income/paystubs/employment/identity
synced_at/errored_at.

To accommodate other payroll aggregators, and better record sync
statuses, let's create a WebhookEvents model to record which webhooks
we've received for a given PayrollAccount. Success or failure, then, is
not determined by the columns, but by the presence of the necessary
WebhookEvents for that payroll aggregator.