Update bin/will-deploy script to handle "FFS-1234-2:" prefix (#485)
Tim's been using prefixes like "FFS-1234-2:" to denote that it's his
second PR addressing a certain ticket. Seems fine to me. Let's update
the deploy script to linkify that syntax as well.

While here, it also makes the Github PR numbers, i.e. the "(#123)", a
link to the PR on GH.

I added some tests to the script as well since those regular expressions
are getting kind of bananas.