Remove `test_fixture_directory` fixture configuration variable
This variable was intended to configure our various request stubs, to
stub based on either the pinwheel or argyle implementations. I find it
to make the rspec test stubbing even more indirect, and since rspec test
stubbing is one of the most complicated aspects of the codebase's
learning curve, I'd prefer we simplify it by just creating pinwheel and
argyle specific test stubs when necessary.