Store `argyle_user_id` on the CbvFlow
We originally had stored this for the prior Argyle implementation, and
just had removed the column (in 18fc68ad0).

The only reason we need to store this is because the
`accounts.connected` webhook doesn't include the `external_id` we used
when creating the user object. Since it doesn't have the `external_id`,
the `user` id is the only other option we could use to tie the incoming
`accounts.connected` webhook to the correct Cbv Flow.

This commit stores the user's id on the CbvFlow, taking care to not
replace it once it already has been stored. This has a weird edge case
when trying to link the same employer twice - we'll probably have to
figure out how to handle this properly so the user doesn't get stuck if
they go back to the employer search after linking the employer once.