Add 1-second artificial delay after Argyle modal closes
This seems necessary so the user doesn't end up on the /synchronizations
page before the Argyle account has been connected.

This is kind of gross, though, and I worry about if Argyle doesn't send
us this webhook in the 1-second timeframe. Since this is probably an
issue with Pinwheel as well, I made FFS-2675 to track updating the
synchronizations page to handle this situation.