clean up CBV flow spec and move client spec to e2e folder for now
the e2e folder doesnt run anywhere so TODO but to start cleaning up cbv flow test to actually pass.

shared examples are mostly an antipattern when it comes to debugging, we should DRY another way if we want. that being said, starting with not thinking about internalization seemed like the stronger start here.