Fix tests running at 23:59pm UTC
Here's a fun one. Because the `Date.current` is evaluated as the file
loads, if the test suite starts up at 23:59pm, any tests that run in
the minute thereafter will be using the improper validation.

By using a proc to load the date range, this can be avoided (so the
validation will use the range accurate as of whenever the validation is
running).

I'm not sure whether this was a bug in production or not. I'd think we
would have heard of it if it were...?