Update PaymentDetailsController spec back to using request stubs
We experimented with stubbing out the PinwheelReport in the
PaymentDetailsController spec, which was a cool idea but never fully
implemented. This spec was the only place we tested some different
Pinwheel response errors, which weren't represented in the stub, and
there were some errors that are happening with the real data as a
result.

* Fix fetching income when the income job fails. (Fetching income will
  also then fail, resulting in an error while processing the report.
  This deserves a more holistic fix but for now this addresses the
  common error case.)