Replace dry-validation with ActiveModel::Validations
Although this library did what Jeff needed it to do, it is
almost entirely duplicative of Rails' built-in validation library
(ActiveModel::Validations).

We do currently use ActiveModel::Validations for frontend error display,
but I think it is less conceptual overhead to have a second use of
ActiveModel::Validation (even if not as conceptually clean) rather than
having to learn a new library to accomplish the same functionality.