FFS-2789: Use Argyle /employer-search endpoint for search
We previously used Argyle's /items endpoint for employer search, but
this returns duplicate results (e.g. for "Allstate").

Argyle customer support suggests using /employer-search to match the
search logic on their end. This is a drop-in replacement because the
endpoints take the same arguments and return the data in the same
format.