FFS-2746: tweak pinwheel account syncher to be more resilient (#579)
the reports aggregator functions presume the data is being fetched for each account, whereas there was no guarantee of that. this attempts a compromise between the two by fetching data for whatever accounts we think are possible.

this also adds redirect behavior if the reports page gets into a funny place--we THINK this isn't possible, but the current redirect safety guards will let us know.

---------

Co-authored-by: Ian Norris <iannorris@Ians-MacBook-Air.local>