FFS-2729: re-link continue button to applicant info page in case they want to continue forward (#591)
Changes
Links the next link on the missing results page such that it would proceed to the application information page, which they will need to fill out if in the LA flow. The application controller already checks if it is unnecessary to be there, so won't hurt non-LA flows while avoiding too much spreading logic.