refactor to begin ability to remove maybe_later
i suspect this might also grant some performance boosts by removing thread burden on our application servers.

this is a refactor with the goal of first behaving exactly as it currently does--putting it into MaybeLater still, but in a way that ActiveJob understands it.

Then if this refactor still behaves as expected, we can move to the next phase --> putting it into asynchronous jobs.

GenericEventTracker itself wasnt very tested, but the controller specs poked at various aspects of it--but can add more unit test coverage of the job/tracker itself now that its responsibilities from the request cycle are way more obvious.