FFS-2754: finish extracting out maybe later (#597)
this removes the usage of the environment variable for job enqueueing, which enables us to remove maybe later.

To fix in a followup PR maybe: right now, if the job fails because it needed to connect to mixpanel or newrelic, the job will fail and retry the whole re-sync. at some point might be worth having a job fire off jobs? (mixpanel and newrelic seperately) [or rip out the newrelic events]


---------

Co-authored-by: Ian Norris <iannorris@Ians-MacBook-Air.local>