FFS-2817: Add NewRelic monitoring role to AWS terraform
We want to be able to ingest AWS metrics into NewRelic so we can set up
dashboards, alerts, and the like. To do so, we need to give NewRelic
readonly access to everything.

We already had a limited form of NewRelic integration in our email
module - I took this time to more properly support configuration of this
parameter via variables in the env config.