bring back libvips because i suspect thats why the background worker started failing (#652)
not that its about libvips in particular, but this is doing the quick fix before taking time to address technical debt later.

why did this break things?

wkhtmltopdf depends on a host of dependencies, and I think libvips's dependencies is secretly saving us here. because wkhtmltopdf is a dragon of a dependency management problem, lets make that better another day.

---------

Co-authored-by: Ian Norris <iannorris@Ians-MacBook-Air.local>