Update `flow_started_seconds_ago` to be relative to consent page
Rather than using `Time.now`, let's use
`consented_to_authorized_use_at`, since it represents when the user
finished the CBV flow. This will allow the calculated value of
`flow_started_seconds_ago` to not be changed if there is a long queue or
some error in processing the CaseWorkerTransmitterJob.