Use `cbv_flow` created_at date as basis for report fetching
It seems to make sense that a report should not include any days of data
after the report was originally linked, even if rendered days later (as
would be the case if there was a transmission failure.)

Let's use the Cbv Flow's creation timestamp for this basis.

Also, add some tests for the report fetching.