FFS-2953: AZ string token translations (#736)
Strings translated for 14.1 and AZ go-live. This PR includes translations for all known English strings; missing Spanish tokens should now be treated as bugs. This incorporates Patricia's feedback on all strings, but Yuleidy has not yet weighed in on the strings we just got back from Lionbridge. There will be a subsequent PR with touchups.