FFS-2937: Fix age range parsing to require YYYY-mm-dd format
We've been seeing an abnormal number of people categorized as "90+".
Upon further inspection, this is due to hiting the `else` in the case
statement when the payroll aggregator returns us bad data (of the form
"--09-04"). Paylocity and Paycor seem to be the culprits.

To handle this, let's return `nil` from the `get_age_range` function
when the date is not of the form "YYYY-mm-dd". This will prevent our
analytics from being polluted.

I also updated the `case` statement to not fall back to "90+" as the
`else` as an extra line of defense against other logic bugs later.