Fix missing "host" ArgumentError on /submits
In dev, we don't always have a value defined for
`client_agency.agency_domain`. To allow the `CbvFlow#to_generic_url`
method to succeed, let's remove the `host` option when the value is
`nil`. This will fall back to
`Rails.application.routes.default_url_options[:host]`, which is always
defined.