FFS-2676: Store link expiration in database field (#780)
* FFS-2676: Store link expiration in database field

Adds the expires_at field to CbvFlowInvitation so that expiration times
can be queried in the database. Because by default Rails stores and
returns this in UTC, also adds #expires_at_local to return in the
configured valid time zone, and adds migrations to create the field and
then populate it for existing records.