Replace template-infra E2E tests with our own
The code we got from Nava's template-infra gives us an `e2e` directory.
Due to the complexity of the integrations used by the Rails app, we need
the simulated browser testing to actually hook into the Rails app more
deeply than would be possible by a separate top-level module invoking
playwright.

If we end up going down a microservice path, maybe there's a world in
which we bring this back to test things holistically (but I don't think
we'd be able to easily do it with the stubs, so it'd have to be live.)