Obscure Pinwheel AWS Key in S3 URLs when saving E2E recordings
This is not strictly necessary, but Github is probably going to continue
flagging these as security alerts.

Note that this is a bit unlike the others since we won't be able to
reconstitute a valid value for this after we redact it. So we'll just
have VCR put back in the placeholder "<PINWHEEL_AWS_KEY_ID_VALUE>" since
we don't use these URLs anyway.