Use expirationless session_store in E2E tests
This is necessary because we will be stubbing the Rails Time.now (the
basis for the cookies' `Expires` attribute), but Chrome will happily
invalidate the cookie based on the real time the test is being run.

To eliminate this issue, this commit replaces the session_store with one
that doesn't expire cookies.