Fix factories not instantiating CbvApplicant STI subclasses
Due to the lack of `initialize_with` in the CbvApplicant factory,
FactoryBot always was instantiating a base CbvApplicant rather than the
desired subclass.

There was one spec failure, in the WeeklyReportMailer spec. Somewhere
deep within ActiveRecord, objects for this spec were getting created
with a `client_agency_id` being produced by the `.sti_name` method. This
method was not symmetric with `.sti_class_for`, as it should be, when
the name has a capital letter within it.

This also makes the same change to our other usage of STI, in
PayrollAccount, although that should be a noop since we don't have any
subclasses of PayrollAccount with capital letters within them.